+++ notifications/__init__.py
@@ -1,0 +1,3 @@
+from notifications.TelegramNotifier import TelegramNotifier
+
+__all__ = ['TelegramNotifier']